feat(1769): transfer the test from webhooks to webhooks.helper
